Update README.md
We simplified the Readme to provide a clearer, more concise notice. The update clarifies that the words come from Gartic's internal database, explains how to report incorrect words to Gartic, and guides contributors to follow CONTRIBUTING.md for further instructions. This change makes the information easier to understand and ensures better communication with users and contributors.